MSVC: Enabled warning (and fixed them) as errors on renderers [skip Travis]

+++ src/Layers/xrRenderDX10/dx10HW.cpp
@@ -439,9 +439,9 @@ void CHW::Present()
     {
         const float fps = Device.GetStats().fFPS;
         if (fps < 30)
-            m_pSwapChain2->SetSourceSize(Device.dwWidth * 0.85f, Device.dwHeight * 0.85f);
+            m_pSwapChain2->SetSourceSize(UINT(Device.dwWidth * 0.85f), UINT(Device.dwHeight * 0.85f));
         else if (fps < 15)
-            m_pSwapChain2->SetSourceSize(Device.dwWidth * 0.7f, Device.dwHeight * 0.7f);
+            m_pSwapChain2->SetSourceSize(UINT(Device.dwWidth * 0.7f), UINT(Device.dwHeight * 0.7f));
     }
 #endif
     CurrentBackBuffer = (CurrentBackBuffer + 1) % BackBufferCount;
